token_healing mode and out constraint mode use chuncked prefill. (#846)
