+++ lightllm/server/router/model_infer/mode_backend/base_backend.py
@@ -84,7 +84,7 @@ class ModeBackend:
         self.disable_chunked_prefill = kvargs.get("disable_chunked_prefill", False)
         self.chunked_prefill_size = kvargs.get("chunked_prefill_size", None)
         self.return_all_prompt_logprobs = kvargs.get("return_all_prompt_logprobs", False)
-        self.use_dynamic_prompt_cache = kvargs.get("use_dynamic_prompt_cache", False)
+        self.use_dynamic_prompt_cache = not kvargs.get("disable_dynamic_prompt_cache", False)
         self.eos_id: List[int] = kvargs.get("eos_id", [2])
         self.disable_cudagraph = kvargs.get("disable_cudagraph", False)
