+++ test/format_out/test_constraint_server.py
@@ -9,8 +9,7 @@ python -m lightllm.server.api_server --model_dir /Meta-Llama-3-8B-Instruct  \
                                      --port 8017                   \
                                      --tp 1                         \
                                      --max_total_token_num 100000 \
-                                     --simple_constraint_mode \
-                                     --use_dynamic_prompt_cache
+                                     --simple_constraint_mode
 """
